src/accounts/login/pyproject.toml
[project]
name = "login"
version = "0.1.0"
description = "Handles logging into an account."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
    "pika>=1.3.2",
    "scylla-driver>=3.28.2",
    "valkey>=6.1.0",
]
